Debug flag
I did not added "verbose" alias because it's incorrect--the "debug" flag
shortens wait time and that's not "verbose" or not. Shortening wait time
is not appropriate for "--debug" too, but never it is verbosity.

The original version also has "-v" alias in flag handling code, but it
won't be handled because "version" flag has "-v" alias and it captures
the arg eariler.